@@ -1,35 +1,36 @@
 source 'http://rubygems.org'
+ruby '1.9.3'
 
-gem 'rails', '3.0.20'
+gem 'rails', '4.1.2'
 
 # Bundle edge Rails instead:
 # gem 'rails', :git => 'git://github.com/rails/rails.git'
 
-gem 'mongoid', "~> 2.0"
-gem 'mongo_ext'
-gem 'mongoid_search', '0.2.4'
-gem 'mongoid_taggable', :git => "git://github.com/wilkerlucio/mongoid_taggable.git"
-gem "mongoid-paperclip", :require => "mongoid_paperclip"
+gem 'mongoid', '~> 4.0'
+gem 'mongoid_search'
+gem 'mongoid_taggable', :git => 'git://github.com/wilkerlucio/mongoid_taggable.git'
+gem 'mongoid-paperclip', :require => 'mongoid_paperclip'
 
-gem 'heroku'
-gem 'foreman'
-gem 'aws-s3', :require => 'aws/s3'
+gem 'heroku', '3.8.4'
+gem 'rake', '10.3.2'
+gem 'aws-sdk'
 gem 'thin'
 
 group :test, :development do
   gem 'rr'
   gem 'hoe', '2.8.0'
-  gem 'test-unit', '=1.2.3'
-  gem "mongoid-rspec"
-  gem 'rspec-core', '2.5.2'
-  gem 'rspec-rails', '~> 2.5.0'
+  gem 'mongoid-rspec'
+  gem 'rspec-core'
+  gem 'rspec-rails'
   gem 'webrat'
-  gem 'cucumber-rails'
+  gem 'cucumber-rails', :require => false
   gem 'cucumber' 
   gem 'capybara'
   gem 'database_cleaner'
   gem 'factory_girl_rails'
   gem 'spork'
-  gem 'rcov'
+  gem 'simplecov'
+  gem 'foreman'
+  gem 'dotenv-rails'
 end
 
