@@ -1,4 +1,4 @@
-AWS::S3::Base.establish_connection!(
+AWS.config(
   :access_key_id     => ENV['S3_KEY'],
   :secret_access_key => ENV['S3_SECRET']
 )
