@@ -1,7 +1,12 @@
 # Be sure to restart your server when you modify this file.
 
-# Your secret key for verifying the integrity of signed cookies.
+# Your secret key is used for verifying the integrity of signed cookies.
 # If you change this key, all old signed cookies will become invalid!
+
 # Make sure the secret is at least 30 characters and all random,
 # no regular words or you'll be exposed to dictionary attacks.
-Lojinha::Application.config.secret_token = 'cc27a5b06f154055c6d43b17e0fd912c4ce9596f10e5ad2c749fb3b8fb513bf41ad944ae2e601b976f58585399ebcd69d8939247e3970e373c8a37378c95da98'
+# You can use `rake secret` to generate a secure secret key.
+
+# Make sure your secret_key_base is kept private
+# if you're sharing your code publicly.
+Lojinha::Application.config.secret_key_base = 'f521e8b4ca69f831127b3f2d6ca2fd750412ca4b957d595a85ae787bd310f25fbee0b58e1746bbe3f95915e3abea8316cc5cf42ddac110b12f6afe13119ea960'
